noop refactor sticky functionality into seperate component /composable bc vue3 convention, no better place
