+++ src/components/Card.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { reactive, computed, onMounted, onBeforeUnmount, onUnmounted, watch, ref, nextTick } from 'vue'
+import { reactive, computed, onMounted, onBeforeUnmount, onUnmounted, watch, ref, toRef, nextTick } from 'vue'
 
 import { useGlobalStore } from '@/stores/useGlobalStore'
 import { useConnectionStore } from '@/stores/useConnectionStore'
@@ -13,6 +13,8 @@ import { useGroupStore } from '@/stores/useGroupStore'
 import { useUploadStore } from '@/stores/useUploadStore'
 import { useBroadcastStore } from '@/stores/useBroadcastStore'
 import { useThemeStore } from '@/stores/useThemeStore'
+
+import { useStickyCard } from '@/composables/useStickyCard.js'
 
 import utils from '@/utils.js'
 import Frames from '@/components/Frames.vue'
@@ -71,13 +73,6 @@ const defaultCardPosition = 100
 const lockingPreDuration = 100 // ms
 const lockingDuration = 100 // ms
 let lockingAnimationTimer, lockingStartTime, shouldCancelLocking
-
-// sticky
-const stickyTimerDuration = 250
-let preventSticking = false
-let stickyTimerComplete = false
-let stickyTimer
-let stickyMap
 
 let observer
 
@@ -159,10 +154,6 @@ const state = reactive({
   isLocking: true,
   lockingPercent: 0,
   lockingAlpha: 0,
-  stickyTranslateX: 0,
-  stickyTranslateY: 0,
-  isAnimationUnsticking: false,
-  stickyStretchResistance: 6,
   defaultBackgroundColor: '#e3e3e3',
   pathIsUpdated: false,
   isVisibleInViewport: false,
@@ -435,7 +426,7 @@ const cardWrapStyle = computed(() => {
     delete styles.zIndex
   }
   if (!globalStore.currentUserIsDraggingCard) {
-    styles.transform = `translate(${state.stickyTranslateX}, ${state.stickyTranslateY})`
+    styles.transform = `translate(${stickyState.stickyTranslateX}, ${stickyState.stickyTranslateY})`
   }
   if (isSnappingToItem.value) {
     styles.opacity = consts.itemSnapOpacity
@@ -1697,161 +1688,13 @@ const shouldNotStick = computed(() => {
     globalStore.currentUserIsResizingList
   )
 })
-const updateShouldNotStickMap = () => {
-  stickyMap = []
-  const element = cardElement.value
-  let rect
-  // connector
-  const connector = element.querySelector('.connector')
-  if (connector) {
-    rect = connector.getBoundingClientRect()
-    rect = utils.rectDimensions(rect)
-    stickyMap.push(rect)
-  }
-  // checkbox
-  const checkbox = element.querySelector('.item-checkbox-button')
-  if (checkbox) {
-    rect = checkbox.getBoundingClientRect()
-    rect = utils.rectDimensions(rect)
-    stickyMap.push(rect)
-  }
-  // tilt resize buttons
-  const tiltResizeButtons = element.querySelectorAll('.bottom-button-wrap')
-  tiltResizeButtons.forEach(button => {
-    rect = button.getBoundingClientRect()
-    rect = utils.rectDimensions(rect)
-    stickyMap.push(rect)
-  })
-}
-const initStickToCursor = () => {
-  preventSticking = false
-  if (shouldNotStick.value || consts.userPrefersReducedMotion()) {
-    preventSticking = true
-  }
-  stickyTimer = setTimeout(() => {
-    stickyTimerComplete = true
-  }, stickyTimerDuration)
-  updateStickyStretchResistance()
-  updateShouldNotStickMap()
-}
-const clearStickyTimer = () => {
-  clearTimeout(stickyTimer)
-  stickyTimerComplete = false
-}
-const stopSticking = () => {
-  clearStickyPositionOffsets()
-  preventSticking = true
-}
-const updateStickyStretchResistance = () => {
-  const zoom = globalStore.getSpaceZoomDecimal
-  let { height, width } = props.card
-  height = height * zoom
-  width = width * zoom
-  // larger sizes have stick less
-  let stretchResistance // higher resistance moves less
-  const area = width * height
-  switch (true) {
-    case area > 1000000:
-      stretchResistance = 24
-      break
-    case area > 60000:
-      stretchResistance = 16
-      break
-    case area > 20000:
-      stretchResistance = 10
-      break
-    default:
-      stretchResistance = 6
-  }
-  state.stickyStretchResistance = stretchResistance
-}
-const stickToCursor = (event) => {
-  if (state.isAnimationUnsticking) { return }
-  if (preventSticking) { return }
-  if (!stickyTimerComplete) { return }
-  const position = utils.cursorPositionInSpace(event)
-  // stop sticking by map
-  const positionIsInsideMap = stickyMap.find(rect => utils.isPointInsideRect(position, rect))
-  if (positionIsInsideMap) {
-    stopSticking()
-    return
-  }
-  // stop sticking by element
-  const classes = ['checkbox-wrap', 'button-wrap', 'progress-wrap', 'inline-button', 'badge']
-  const elements = ['button', 'progress', 'iframe']
-  const isOverAction = classes.includes(event.target.className) || elements.includes(event.target.nodeName.toLowerCase())
-  const isOverTag = event.target.className.includes('button-badge')
-  if (shouldNotStick.value || isOverAction || isOverTag) {
-    stopSticking()
-    return
-  }
-  const isButtonHover = event.target.closest('.inline-button-wrap') || event.target.closest('.button-wrap')
-  if (isButtonHover) {
-    stopSticking()
-    return
-  }
-  // position
-  const stretchResistance = state.stickyStretchResistance
-  const { height, width } = props.card
-  const halfWidth = width / 2
-  const halfHeight = height / 2
-  const centerX = x.value + halfWidth
-  const centerY = y.value + halfHeight
-  // position from card center
-  const xFromCenter = position.x - centerX
-  const yFromCenter = position.y - centerY
-  // percentage from center to card edge
-  const xPercent = (xFromCenter / halfWidth)
-  const yPercent = (yFromCenter / halfHeight)
-  // calc sticky offset
-  let xOffset = (xPercent * halfWidth) / stretchResistance
-  xOffset = Math.round(xOffset)
-  let yOffset = (yPercent * halfHeight) / stretchResistance
-  yOffset = Math.round(yOffset)
-  state.stickyTranslateX = xOffset + 'px'
-  state.stickyTranslateY = yOffset + 'px'
-}
-const unstickToCursor = () => {
-  clearStickyTimer()
-  state.isAnimationUnsticking = true
-  const xOffset = parseInt(state.stickyTranslateX)
-  const yOffset = parseInt(state.stickyTranslateY)
-  const timing = {
-    duration: 0, // sum of keyframe offsets
-    easing: 'cubic-bezier(0.45, 0, 0.55, 1)',
-    iterations: 1
-  }
-  const swings = [-0.9, 0.6, -0.4, 0.2, 0] // [-1, 0.75, -0.5, 0.25, 0]
-  let keyframes = [
-    { transform: `translate(${xOffset * swings[0]}px,   ${yOffset * swings[0]}px)`, offset: 50 },
-    { transform: `translate(${xOffset * swings[1]}px, ${yOffset * swings[1]}px)`, offset: 75 },
-    { transform: `translate(${xOffset * swings[2]}px, ${yOffset * swings[2]}px)`, offset: 50 },
-    { transform: `translate(${xOffset * swings[3]}px, ${yOffset * swings[3]}px)`, offset: 100 },
-    { transform: `translate(${xOffset * swings[4]}px,    ${yOffset * swings[4]}px)`, offset: 100 }
-  ]
-  keyframes.forEach(keyframe => {
-    timing.duration = timing.duration + keyframe.offset
-  })
-  let lastOffset = 0
-  keyframes = keyframes.map(keyframe => {
-    keyframe.offset = lastOffset + (keyframe.offset / timing.duration)
-    keyframe.offset = utils.roundFloat(keyframe.offset)
-    lastOffset = keyframe.offset
-    return keyframe
-  })
-  // play animation
-  const element = cardElement.value
-  if (!element) { return }
-  const animation = element.animate(keyframes, timing)
-  animation.onfinish = () => {
-    clearStickyPositionOffsets()
-    state.isAnimationUnsticking = false
-  }
-}
-const clearStickyPositionOffsets = () => {
-  state.stickyTranslateX = 0
-  state.stickyTranslateY = 0
-}
+const { stickyState, initStickToCursor, stickToCursor, unstickToCursor, clearStickyPositionOffsets } = useStickyCard({
+  card: toRef(props, 'card'),
+  cardElement,
+  x,
+  y,
+  shouldNotStick
+})
 
 // lock card position
 
@@ -2038,7 +1881,7 @@ const toggleVideoIsPaused = () => {
   :data-should-render="shouldRender"
   :data-is-locked="isLocked"
   :data-tilt-degrees="card.tilt"
-  :data-sticky-stretch-resistance="state.stickyStretchResistance"
+  :data-sticky-stretch-resistance="stickyState.stickyStretchResistance"
   :data-x="x"
   :data-y="y"
   :data-card-x="props.card.x"
